refactor: cleanup code
