+++ src/market/proposal/index.ts
@@ -1,4 +1,3 @@
-export * from "./types";
 export * from "./market-proposal-event";
 export * from "./market-proposal";
 export * from "./offer-proposal";
